# platform-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user/create": {
            "post": {
                "description": "UserCreate Api 创建用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserCreate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create User request body",
                        "name": "UserCreate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user/delete": {
            "delete": {
                "description": "UserDelete Api 删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Delete User request body",
                        "name": "UserDelete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/user/list": {
            "get": {
                "description": "UserList Api 查询用户的信息，包含用户权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页索引",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserListRes"
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "UserLogIn Api 用户登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserLogIn",
                "parameters": [
                    {
                        "description": "UserLogin request body",
                        "name": "UserLogin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginApiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginApiRes"
                        }
                    }
                }
            }
        },
...
        "/api/user/refresh_token": {
            "post": {
                "description": "UserRefreshToken Api 刷新access-token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserRefreshToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "UserRefreshToken request body",
                        "name": "UserRefreshToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRefreshTokenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserRefreshTokenRes"
                        }
                    }
                }
            }
        },
        "/api/user/reset_password": {
            "post": {
                "description": "UserResetPassword Api 重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserResetPassword",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Reset User Password request body",
                        "name": "UserResetPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserResetPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user/update_info": {
            "put": {
                "description": "UserUpdateInfo Api 更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserUpdateInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update User Info request body",
                        "name": "UserUpdateInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserUpdateInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user/update_password": {
            "post": {
                "description": "UserUpdatePassword Api 更新密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "UserUpdatePassword",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "UserUpdatePassword request body",
                        "name": "UserUpdatePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserUpdatePasswordApiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.Permission": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "子资源，可能为空",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Permission"
                    }
                },
                "key": {
                    "$ref": "#/definitions/enum.PermissionKey"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserCreateReq": {
            "type": "object",
            "required": [
                "user_display_name",
                "user_name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_display_name": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "dto.UserDeleteReq": {
            "type": "object",
            "required": [
                "user_ids"
            ],
            "properties": {
                "user_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "dto.UserListItem": {
            "type": "object",
            "properties": {
                "create_user_display_name": {
                    "type": "string"
                },
                "create_user_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Permission"
                    }
                },
                "phone_number": {
                    "type": "string"
                },
                "role_infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserRoleInfo"
                    }
                },
                "user_display_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "dto.UserListRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserListItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.UserLoginApiReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Use DES-DCB encryption",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UserLoginApiRes": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserRefreshTokenReq": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "dto.UserRefreshTokenRes": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "dto.UserResetPasswordReq": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserRoleInfo": {
            "type": "object",
            "properties": {
                "role_id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserUpdateInfoReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_display_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.UserUpdatePasswordApiReq": {
            "type": "object",
            "required": [
                "confirm_password",
                "new_password",
                "old_password"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "description": "Use DES-DCB encryption",
                    "type": "string"
                }
            }
        },
...
        "enum.PermissionKey": {
            "type": "string",
            "enum": [
                "view",
                "edit",
                "delete",
                "poweron",
                "poweroff",
                "permissionmgmt"
            ],
            "x-enum-varnames": [
                "PermissionKeyView",
                "PermissionKeyEdit",
                "PermissionKeyDelete",
                "PermissionKeyPowerOn",
                "PermissionKeyPowerOff",
                "PermissionKeyPermissionMgmt"
            ]
        },
...
    }
}